requirements.txt
cyborgdb
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
pydantic
faker
fhir.resources
presidio-analyzer
presidio-anonymizer
spacy
prefect
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
torch
transformers
sentence-transformers
numpy
cryptography
minio
python-jose[cryptography]
passlib[bcrypt]
slowapi
python-multipart
# New authentication dependencies
bcrypt>=4.0.1
PyJWT>=2.8.0
requests>=2.31.0
pyotp>=2.9.0
